# internal/http/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "rssam API",
    "version": "1.0.0",
    "description": "REST API of the rssam RSS aggregator.\n\n**Authentication.** Every `/v1` endpoint takes `X-Auth-Token: <token>` where the token is either a personal API key (`POST /v1/me/api-keys`) or the global `AUTH_TOKEN` (maps to user id 1). A browser session cookie is also accepted, but state-changing requests authenticated only by the cookie must pass the same-origin check (`Sec-Fetch-Site`/`Origin`), otherwise 403.\n\n**Envelope.** Successful JSON responses are `{\"data\": ..., \"total\": n}`; errors are `{\"error_message\": \"...\"}`. Lists take `limit`/`offset`; `total` is the full row count.\n\n**Ownership.** Objects are scoped to the calling user; reading another user's object yields 404. Endpoints marked \"Requires an admin principal\" return 403 for non-admins.\n\n**Request id.** Every response carries `X-Request-Id` (echoed from the request when a proxy set it, generated otherwise); quote it when reporting 5xx errors.\n\n**Limits.** Request bodies are capped (MAX_REQUEST_BODY_BYTES, default 1 MiB) → 413; per-IP rate limits on authentication failures and sensitive routes → 429."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "security": [
    {
      "AuthToken": []
    },
    {
      "SessionCookie": []
    }
  ],
  "paths": {
...
    "/v1/feeds/{feedID}/entries/{entryID}": {
      "parameters": [
        {
          "name": "feedID",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "description": "Feed id."
        },
        {
          "name": "entryID",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "description": "Entry id."
        }
      ],
      "get": {
        "tags": [
          "entries"
        ],
        "summary": "Get entry of a feed",
        "operationId": "getFeedEntry",
        "responses": {
          "200": {
            "description": "Entry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "put": {
        "tags": [
          "entries"
        ],
        "summary": "Update entry status/starred",
        "operationId": "updateFeedEntry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/entries": {
      "get": {
        "tags": [
          "entries"
        ],
        "summary": "List entries",
        "operationId": "listEntries",
        "parameters": [
          {
            "$ref": "#/components/parameters/limit"
          },
          {
            "$ref": "#/components/parameters/offset"
          },
          {
            "$ref": "#/components/parameters/entryStatus"
          },
          {
            "$ref": "#/components/parameters/entryStarred"
          },
          {
            "$ref": "#/components/parameters/entrySort"
          },
          {
            "$ref": "#/components/parameters/entrySearch"
          },
          {
            "$ref": "#/components/parameters/feedIDQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Entries across all feeds of the caller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "put": {
        "tags": [
          "entries"
        ],
        "summary": "Bulk update entries",
        "operationId": "bulkUpdateEntries",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntriesBulkUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Number of rows updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkUpdateResultResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/v1/entries/{id}/fetch-content": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "description": "Object id."
        }
      ],
      "get": {
        "tags": [
          "entries"
        ],
        "summary": "Fetch full article content",
        "operationId": "fetchEntryContent",
        "description": "Downloads the entry URL (SSRF-guarded), applies the feed's scraper rules or the readability extractor, and stores the result.",
        "responses": {
          "200": {
            "description": "Entry with content replaced by the extracted article; original kept in original_content.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "502": {
            "description": "The article page could not be fetched or parsed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Token",
        "description": "Personal API key or the global AUTH_TOKEN."
      },
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_id",
        "description": "Web UI session. Mutations additionally require a same-origin request."
      },
      "MetricsBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "METRICS_TOKEN for /metrics only."
      }
    },
    "parameters": {
      "limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "maximum": 10000,
          "default": 100
        },
        "description": "Page size; values above 10000 are clamped."
      },
      "offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 0,
          "default": 0
        }
      },
      "entryStatus": {
        "name": "status",
        "in": "query",
        "schema": {
          "$ref": "#/components/schemas/EntryStatus"
        }
      },
      "entryStarred": {
        "name": "starred",
        "in": "query",
        "schema": {
          "type": "boolean"
        }
      },
      "entrySort": {
        "name": "sort",
        "in": "query",
        "schema": {
          "type": "string",
          "enum": [
            "newest",
            "oldest"
          ],
          "default": "newest"
        }
      },
      "entrySearch": {
        "name": "q",
        "in": "query",
        "schema": {
          "type": "string"
        },
        "description": "Full-text search over title and content."
      },
      "feedIDQuery": {
        "name": "feed_id",
        "in": "query",
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed input or validation error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Object not found or not owned by the caller.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit exceeded.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected server error. The response carries an X-Request-Id header to correlate with server logs.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "The backing store for this resource is not configured.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error_message": {
            "type": "string"
          }
        },
        "required": [
          "error_message"
        ],
        "description": "Error envelope returned by every non-2xx JSON response."
      },
...
      "Enclosure": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "mime_type": {
            "type": "string"
          }
        },
        "required": [
          "url",
          "size",
          "mime_type"
        ]
      },
      "Entry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "feed_id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "content": {
            "type": "string"
          },
          "original_content": {
            "type": "string",
            "description": "Feed content kept when full content was fetched."
          },
          "content_fetched": {
            "type": "boolean"
          },
          "author": {
            "type": "string"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "hash": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "starred": {
            "type": "boolean"
          },
          "enclosures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Enclosure"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "feed_id",
          "title",
          "url",
          "content",
          "content_fetched",
          "hash",
          "status",
          "starred",
          "created_at",
          "updated_at"
        ]
      },
      "EntryStatus": {
        "type": "string",
        "enum": [
          "unread",
          "read",
          "removed"
        ]
      },
      "EntryUpdate": {
        "type": "object",
        "properties": {
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "starred": {
            "type": "boolean"
          }
        },
        "description": "At least one field must be present."
      },
      "EntriesBulkUpdate": {
        "type": "object",
        "properties": {
          "entry_ids": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "status": {
            "$ref": "#/components/schemas/EntryStatus"
          },
          "starred": {
            "type": "boolean"
          }
        },
        "required": [
          "entry_ids"
        ],
        "description": "entry_ids: 1..1000 ids owned by the caller. Set status and/or starred."
      },
      "BulkUpdateResult": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "integer"
          }
        },
        "required": [
          "updated"
        ]
      },
...
      "EntryList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Entry"
            }
          },
          "total": {
            "type": "integer",
            "description": "Number of items in data (1 for single objects) or total rows for lists."
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
...
      "EntryResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Entry"
          },
          "total": {
            "type": "integer",
            "description": "Number of items in data (1 for single objects) or total rows for lists."
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
...
      "BulkUpdateResultResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BulkUpdateResult"
          },
          "total": {
            "type": "integer",
            "description": "Number of items in data (1 for single objects) or total rows for lists."
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
...
    }
  }
}